pyproject.toml
[tool.poetry]
name = "Scripts"
version = "0.1.0"
description = "Scripts de integración de entrono de desarrollo con Bspwm"
authors = ["medicendav <davidtorrezreyes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.0"
beautifulsoup4 = "^4.9.3"
notify2 = "^0.3.1"
dbus-python = "^1.2.16"
pulsectl = "^20.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flakehell = "^0.7.1"
isort = "^5.6.4"
mypy = "^0.790"
rope = "^0.18.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']

[tool.flakehell]
max_line_length = 79
show-source = true
format = "grouped"

[tool.flakehell.plugins]
pyflakes = ["+*", "-E501"]
flake8-blind-except = ["+*"]
flake8-builtins = ["+*"]
flake8-import-order-flake8-linter = ["+*"]
mccabe = ["+*"]
pycodestyle = ["+*"]
